fix: harden owner location classification
