+++ docs/SCRIPT_BEST_PRACTICES.md
@@ -1594,7 +1594,10 @@ Concrete example (owner-housing-distances script):
 ## References
 
 - Reference scripts:
+  - `analytics/dagster/scripts/owner-housing-distances/calculate_distances.py`
+    (canonical implementation)
   - `server/src/scripts/owner-housing-distances/calculate_distances.py`
+    (compatibility entrypoint)
   - `server/src/scripts/owner-housing-distances/test_calculate_distances.py` (tests)
   - `server/src/scripts/import-dpe/import-dpe.py`
